My first Sinatra app
To run, say

  $ ruby hello.rb

The first line is a Ruby require statement that pulls in all
the code from the Sinatra library; it’s needed in all Sinatra apps.
The next block of code is what’s known as a route handler. It starts
with the HTTP verb get, and says that the client should get the page
with the URL of /hello. The last line of the handler is always evaluated
and sent to the browser; in this case it was the string Hello Sinatra!

Signed-off-by: Esa Koponen <esa.koponen@jotautomation.com>